Changes made months ago which should not be lost

+++ src/main/java/com/github/nikolasfunction/tournamentobserver/database/repository/IUserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.github.nikolasfunction.tournamentobserver.database.entity.User;
 
 @Repository
-public interface IUserRepository extends JpaRepository<User, Long> {
+public interface IUserRepository extends JpaRepository<User, Integer> {
 
 }
